corrige tp integrador

+++ POO-TrabajoIntegrador/src/Anfitrion.java
@@ -1,6 +1,6 @@
 public class Anfitrion extends Usuario {
 
-    public Anfitrion (String nombre, String correoElectronico, double dni) {
+    public Anfitrion (String nombre, String correoElectronico, int dni) {
         super(nombre, correoElectronico, dni);
     }
